requirements.txt
Django==3.2.0
psycopg2-binary==2.9.1
requests==2.26.0
Pillow==8.4.0
django-crispy-forms==1.12.0
djangorestframework==3.12.4
django-import-export==3.3.7

